fix gallery
