+++ src/main.js
@@ -6,7 +6,5 @@ require('./game/game');
 require('./getReviews/getReviews');
 require('./getReviews/getReviewBlock');
 require('./reviews');
-require('./galleryModules/getImageSrc');
-require('./galleryModules/showImage');
 require('./gallery');
